+++ zkml-jolt-core/src/jolt/mod.rs
@@ -1,4 +1,420 @@
+//! A state-of-the-art zkVM, called Jolt, which turns almost everything a VM does into reads and writes to memory.
+//! This includes the “fetch-decode-execute” logic of the VM.
+pub mod bytecode;
 pub mod instruction;
+pub mod instruction_lookups;
 pub mod lookup_trace;
+pub mod r1cs;
+pub mod registers;
 pub mod witness;
-pub mod zkvm;
+
+use crate::jolt::{
+    bytecode::{BytecodePreprocessing, BytecodeProof},
+    instruction_lookups::LookupsProof,
+    r1cs::{
+        constraints::{JoltONNXConstraints, R1CSConstraints},
+        spartan::UniformSpartanProof,
+    },
+    registers::RegistersTwistProof,
+};
+use jolt_core::{
+    field::JoltField,
+    poly::{
+        commitment::{commitment_scheme::CommitmentScheme, dory::DoryGlobals},
+        opening_proof::{ProverOpeningAccumulator, VerifierOpeningAccumulator},
+    },
+    utils::{errors::ProofVerifyError, transcript::Transcript},
+};
+use lookup_trace::WORD_SIZE;
+use onnx_tracer::trace_types::{ONNXCycle, ONNXInstr};
+use serde::{Deserialize, Serialize};
+use std::marker::PhantomData;
+
+#[derive(Clone, Serialize, Deserialize)]
+pub struct JoltProverPreprocessing<F, PCS, ProofTranscript>
+where
+    F: JoltField,
+    PCS: CommitmentScheme<ProofTranscript, Field = F>,
+    ProofTranscript: Transcript,
+{
+    pub shared: JoltSharedPreprocessing,
+    field: F::SmallValueLookupTables,
+    _p: PhantomData<(ProofTranscript, PCS)>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct JoltSharedPreprocessing {
+    pub bytecode: BytecodePreprocessing,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct JoltVerifierPreprocessing<F, PCS, ProofTranscript>
+where
+    F: JoltField,
+    PCS: CommitmentScheme<ProofTranscript, Field = F>,
+    ProofTranscript: Transcript,
+{
+    pub shared: JoltSharedPreprocessing,
+    _p: PhantomData<(F, PCS, ProofTranscript)>,
+}
+
+impl<F, PCS, ProofTranscript> From<&JoltProverPreprocessing<F, PCS, ProofTranscript>>
+    for JoltVerifierPreprocessing<F, PCS, ProofTranscript>
+where
+    F: JoltField,
+    PCS: CommitmentScheme<ProofTranscript, Field = F>,
+    ProofTranscript: Transcript,
+{
+    fn from(preprocessing: &JoltProverPreprocessing<F, PCS, ProofTranscript>) -> Self {
+        JoltVerifierPreprocessing {
+            shared: preprocessing.shared.clone(),
+            _p: PhantomData,
+        }
+    }
+}
+
+pub struct JoltSNARK<F, PCS, ProofTranscript>
+where
+    ProofTranscript: Transcript,
+    PCS: CommitmentScheme<ProofTranscript, Field = F>,
+    F: JoltField,
+{
+    pub trace_length: usize,
+    bytecode: BytecodeProof<F, ProofTranscript>,
+    instruction_lookups: LookupsProof<WORD_SIZE, F, PCS, ProofTranscript>,
+    registers: RegistersTwistProof<F, ProofTranscript>,
+    r1cs: UniformSpartanProof<F, ProofTranscript>,
+    _p: PhantomData<PCS>,
+}
+
+impl<F, PCS, ProofTranscript> JoltSNARK<F, PCS, ProofTranscript>
+where
+    ProofTranscript: Transcript,
+    PCS: CommitmentScheme<ProofTranscript, Field = F>,
+    F: JoltField,
+{
+    #[tracing::instrument(skip_all, name = "Jolt::preprocess")]
+    fn shared_preprocess(bytecode: Vec<ONNXInstr>) -> JoltSharedPreprocessing {
+        let bytecode_preprocessing = BytecodePreprocessing::preprocess(bytecode);
+        JoltSharedPreprocessing {
+            bytecode: bytecode_preprocessing,
+        }
+    }
+
+    #[tracing::instrument(skip_all, name = "Jolt::preprocess")]
+    fn prover_preprocess(
+        bytecode: Vec<ONNXInstr>,
+    ) -> JoltProverPreprocessing<F, PCS, ProofTranscript> {
+        let small_value_lookup_tables = F::compute_lookup_tables();
+        F::initialize_lookup_tables(small_value_lookup_tables.clone());
+        let shared = Self::shared_preprocess(bytecode);
+        JoltProverPreprocessing {
+            shared,
+            field: small_value_lookup_tables,
+            _p: PhantomData,
+        }
+    }
+
+    #[tracing::instrument(skip_all, name = "Jolt::prove")]
+    pub fn prove(
+        mut preprocessing: JoltProverPreprocessing<F, PCS, ProofTranscript>,
+        mut trace: Vec<ONNXCycle>,
+    ) -> Self {
+        let trace_length = trace.len();
+        println!("Trace length: {trace_length}");
+        F::initialize_lookup_tables(std::mem::take(&mut preprocessing.field));
+        // pad trace to the next power of two
+        trace.resize(trace.len().next_power_of_two(), ONNXCycle::no_op());
+        let padded_trace_length = trace.len();
+        let ram_addresses: Vec<usize> = trace.iter().map(|cycle| cycle.td() + 1).collect();
+        let ram_K = ram_addresses.iter().max().unwrap().next_power_of_two();
+        let K = [
+            preprocessing.shared.bytecode.code_size,
+            ram_K,
+            1 << 16, // K for instruction lookups Shout
+        ]
+        .into_iter()
+        .max()
+        .unwrap();
+        println!("T = {padded_trace_length}, K = {K}");
+        let _guard = DoryGlobals::initialize(K, padded_trace_length);
+        let mut transcript = ProofTranscript::new(b"Jolt transcript");
+        let mut opening_accumulator: ProverOpeningAccumulator<F, PCS, ProofTranscript> =
+            ProverOpeningAccumulator::new();
+        let constraint_builder = JoltONNXConstraints::construct_constraints(padded_trace_length);
+        let spartan_key = UniformSpartanProof::<F, ProofTranscript>::setup(
+            &constraint_builder,
+            padded_trace_length,
+        );
+        transcript.append_scalar(&spartan_key.vk_digest);
+        let r1cs = UniformSpartanProof::prove::<PCS>(
+            &preprocessing,
+            &constraint_builder,
+            &spartan_key,
+            &trace,
+            &mut transcript,
+        )
+        .ok()
+        .unwrap();
+        let instruction_proof: LookupsProof<WORD_SIZE, F, PCS, ProofTranscript> =
+            LookupsProof::prove(
+                &preprocessing,
+                &trace,
+                &mut opening_accumulator,
+                &mut transcript,
+            );
+        let registers_proof = RegistersTwistProof::prove(
+            &preprocessing,
+            &trace,
+            ram_K,
+            &mut opening_accumulator,
+            &mut transcript,
+        );
+        let bytecode_proof =
+            BytecodeProof::prove(&preprocessing.shared.bytecode, &trace, &mut transcript);
+        JoltSNARK {
+            trace_length,
+            r1cs,
+            registers: registers_proof,
+            instruction_lookups: instruction_proof,
+            bytecode: bytecode_proof,
+            _p: PhantomData,
+        }
+    }
+
+    #[tracing::instrument(skip_all)]
+    pub fn verify(
+        &self,
+        preprocessing: JoltVerifierPreprocessing<F, PCS, ProofTranscript>,
+    ) -> Result<(), ProofVerifyError> {
+        let mut transcript = ProofTranscript::new(b"Jolt transcript");
+        let mut opening_accumulator: VerifierOpeningAccumulator<F, PCS, ProofTranscript> =
+            VerifierOpeningAccumulator::new();
+        // Regenerate the uniform Spartan key
+        let padded_trace_length = self.trace_length.next_power_of_two();
+        let r1cs_builder = JoltONNXConstraints::construct_constraints(padded_trace_length);
+        let spartan_key =
+            UniformSpartanProof::<F, ProofTranscript>::setup(&r1cs_builder, padded_trace_length);
+        transcript.append_scalar(&spartan_key.vk_digest);
+        self.r1cs
+            .verify::<PCS>(&spartan_key, &mut transcript)
+            .map_err(|e| ProofVerifyError::SpartanError(e.to_string()))?;
+        self.instruction_lookups
+            .verify(&mut opening_accumulator, &mut transcript)?;
+        self.registers.verify(
+            padded_trace_length,
+            &mut opening_accumulator,
+            &mut transcript,
+        )?;
+        self.bytecode.verify(
+            &preprocessing.shared.bytecode,
+            self.trace_length,
+            &mut transcript,
+        )?;
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod e2e_tests {
+
+    use crate::{
+        jolt::{JoltProverPreprocessing, JoltSNARK},
+        program::ONNXProgram,
+    };
+    use ark_bn254::Fr;
+    use jolt_core::poly::commitment::dory::DoryCommitmentScheme;
+    use jolt_core::utils::transcript::KeccakTranscript;
+    use log::info;
+    use onnx_tracer::{custom_addsubmul_model, logger::init_logger, model, tensor::Tensor};
+    use serde_json::Value;
+    use std::{collections::HashMap, fs::File, io::Read};
+
+    type PCS = DoryCommitmentScheme<KeccakTranscript>;
+
+    #[test]
+    fn test_custom_addsubmul() {
+        // --- Preprocessing ---
+        let custom_addsubmul_model = custom_addsubmul_model();
+        let program_bytecode = onnx_tracer::decode_model(custom_addsubmul_model.clone());
+        println!("Program code: {program_bytecode:#?}");
+        let pp: JoltProverPreprocessing<Fr, PCS, KeccakTranscript> =
+            JoltSNARK::prover_preprocess(program_bytecode);
+
+        // --- Proving ---
+        let input = Tensor::new(Some(&[60]), &[1]).unwrap();
+        let execution_trace = onnx_tracer::execution_trace(custom_addsubmul_model, &input);
+        println!("Execution trace: {execution_trace:#?}");
+        let snark: JoltSNARK<Fr, PCS, KeccakTranscript> =
+            JoltSNARK::prove(pp.clone(), execution_trace);
+
+        // --- Verification ---
+        snark.verify((&pp).into()).unwrap();
+    }
+
+    /// Load vocab.json into HashMap<String, (usize, i32)>
+    pub fn load_vocab(
+        path: &str,
+    ) -> Result<HashMap<String, (usize, i32)>, Box<dyn std::error::Error>> {
+        let mut file = File::open(path)?;
+        let mut contents = String::new();
+        file.read_to_string(&mut contents)?;
+
+        let json_value: Value = serde_json::from_str(&contents)?;
+        let mut vocab = HashMap::new();
+
+        if let Value::Object(map) = json_value {
+            for (word, data) in map {
+                if let (Some(index), Some(idf)) = (
+                    data.get("index").and_then(|v| v.as_u64()),
+                    data.get("idf").and_then(|v| v.as_i64()),
+                ) {
+                    vocab.insert(word, (index as usize, idf as i32));
+                }
+            }
+        }
+
+        Ok(vocab)
+    }
+
+    /// Tokenize and convert text to vector of length 1000
+    pub fn build_input_vector(text: &str, vocab: &HashMap<String, (usize, i32)>) -> Vec<i128> {
+        let mut vec = vec![0; 1000];
+
+        // Split text into tokens (preserve punctuation as tokens)
+        let re = regex::Regex::new(r"\w+|[^\w\s]").unwrap();
+        for cap in re.captures_iter(text) {
+            let token = cap.get(0).unwrap().as_str().to_lowercase();
+            if let Some(&(index, idf)) = vocab.get(&token) {
+                if index < 1000 {
+                    vec[index] += idf as i128; // accumulate idf value
+                }
+            }
+        }
+
+        vec
+    }
+
+    #[test]
+
+    pub fn test_article_classification_output() {
+        init_logger();
+        let working_dir: &str = "../onnx-tracer/models/article_classification/";
+
+        // Load the vocab mapping from JSON
+        let vocab_path = format!("{working_dir}/vocab.json",);
+        let vocab = load_vocab(&vocab_path).expect("Failed to load vocab");
+
+        // Input text string to classify
+        let input_texts = [
+            "The government plans new trade policies.",
+            "The latest computer model has impressive features.",
+            "The football match ended in a thrilling draw.",
+            "The new movie has received rave reviews from critics.",
+            "The stock market saw a significant drop today.",
+        ];
+
+        let expected_classes = ["politics", "tech", "sport", "entertainment", "business"];
+
+        let mut predicted_classes = Vec::new();
+
+        for input_text in &input_texts {
+            // Build input vector from the input text
+            let input_vector = build_input_vector(input_text, &vocab);
+
+            // Prepare ONNX program
+            let text_classification = ONNXProgram {
+                model_path: format!("{working_dir}network.onnx").into(),
+                inputs: Tensor::new(Some(&input_vector), &[1, 1000]).unwrap(),
+            };
+
+            // Decode to program bytecode (for EZKL use)
+            let program_bytecode = text_classification.decode();
+            println!("Program code: {program_bytecode:#?}");
+
+            // Load model
+            let model = model(&text_classification.model_path);
+
+            // Run inference
+            let result = model
+                .forward(&[text_classification.inputs.clone()])
+                .unwrap();
+            let output = result.outputs[0].clone();
+
+            // Map index to label
+            let classes = ["business", "entertainment", "politics", "sport", "tech"];
+            let (pred_idx, _) = output
+                .iter()
+                .enumerate()
+                .max_by(|a, b| a.1.partial_cmp(b.1).unwrap())
+                .unwrap();
+
+            info!("Output: {}", output.show());
+            info!("Predicted class: {}", classes[pred_idx]);
+
+            predicted_classes.push(classes[pred_idx]);
+        }
+        // Check if predicted classes match expected classes
+        for (predicted, expected) in predicted_classes.iter().zip(expected_classes.iter()) {
+            assert_eq!(predicted, expected, "Mismatch in predicted class");
+        }
+    }
+
+    #[test]
+    fn test_medium_classification() {
+        init_logger();
+        let mut input_vector = vec![846, 3, 195, 4, 374, 14, 259];
+        input_vector.resize(100, 0); // Resize to match the input shape
+
+        let text_classification = ONNXProgram {
+            model_path: "../onnx-tracer/models/medium_text_classification/network.onnx".into(),
+            inputs: Tensor::new(Some(&input_vector), &[1, 100]).unwrap(), // Example input
+        };
+        let program_bytecode = text_classification.decode();
+        println!("Program code: {program_bytecode:#?}",);
+        text_classification.trace();
+    }
+
+    #[test]
+    fn test_medium_classification_output() {
+        init_logger();
+        let mut input_vector = vec![197, 10, 862, 8, 23, 53, 2, 319, 34, 122, 100, 53, 33];
+        input_vector.resize(100, 0); // Resize to match the input shape
+
+        let text_classification = ONNXProgram {
+            model_path: "../onnx-tracer/models/medium_text_classification/network.onnx".into(),
+            inputs: Tensor::new(Some(&input_vector), &[1, 100]).unwrap(), // Example input
+        };
+        let program_bytecode = text_classification.decode();
+        println!("Program code: {program_bytecode:#?}",);
+        let model = model(&text_classification.model_path);
+
+        let result = model
+            .forward(&[text_classification.inputs.clone()])
+            .unwrap();
+        let output = result.outputs[0].clone();
+        info!("Output: {output:#?}",);
+    }
+
+    #[test]
+    fn test_subgraph() {
+        init_logger();
+        let subgraph_program = ONNXProgram {
+            model_path: "../onnx-tracer/models/subgraph/network.onnx".into(),
+            inputs: Tensor::new(Some(&[1, 2, 3, 4]), &[1, 4]).unwrap(), // Example input
+        };
+        let program_bytecode = subgraph_program.decode();
+
+        println!("Program decoded");
+        println!("Program code: {program_bytecode:#?}",);
+
+        // Test that the addresses of a subgraph are monotonically increasing
+        let mut i = 0;
+        for instr in program_bytecode {
+            assert!(instr.address > i);
+            i = instr.address;
+        }
+
+        subgraph_program.trace();
+    }
+}
