+++ zkml-jolt-core/src/jolt/r1cs/constraints.rs
@@ -1,7 +1,7 @@
 use jolt_core::field::JoltField;
 use onnx_tracer::trace_types::CircuitFlags;
 
-use crate::jolt::zkvm::r1cs::{
+use crate::jolt::r1cs::{
     builder::{CombinedUniformBuilder, R1CSBuilder},
     inputs::JoltONNXR1CSInputs,
 };
