+++ zkml-jolt-core/src/jolt/r1cs/inputs.rs
@@ -5,9 +5,9 @@
 )]
 
 use super::spartan::UniformSpartanProof;
+use crate::jolt::JoltProverPreprocessing;
 use crate::jolt::lookup_trace::LookupTrace;
 use crate::jolt::witness::CommittedPolynomials;
-use crate::jolt::zkvm::JoltProverPreprocessing;
 use jolt_core::field::JoltField;
 use jolt_core::jolt::instruction::LookupQuery;
 use jolt_core::poly::commitment::commitment_scheme::CommitmentScheme;
